Move astro_coords.py to types/coordinates.py
