+++ python/PiFinder/types/coordinates.py
